auth-server/src/requirements.txt
fastapi
flask
pydantic
uvicorn
pipenv
couchbase
prometheus_client
python-dotenv
starlette
starlette_prometheus
python-multipart